Hacked a web application for Boot Upgrade report
A web application decoupled from the shell allows scanning a dir and creates a report from asciidoc and provides it as web page.
From this report requests can be sent to run single recipes to remediate changes. After each recipe a new report is created allowing to "check off" the list of changes.