Reimplementation of OpenRewrite AST parser for Maven
Two implementations currently exist:
- RewriteMavenProjectParser
- RewriteProjectParser

RewriteMavenProjectParser
creates Maven Plexus container to retrieve the Maven instance.
A Maven instance is used to call `clean` and `package` while registering a listener
to the lifecycle event when the goals are finished.
The MavenSession provided to the listener allows executing the same OR classes as used when
running in the rewrite-maven plugin.
The created AST should be exactly the same as with the Maven plugin.

RewriteProjectParser
replicates the RewriteMavenProjectParser while split into smaller components.
The RewriteMavenProjectParser is used to create the target AST which should be
identically when parsing the same project with RewriteProjectParser.
The RewriteProjectParser is meant to replace the RewriteMavenProjectParser.